requirements.txt
aiohttp
pandas
numpy
python-dotenv
cryptography
requests
